% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/capital_forecast.R
\name{CapitalForecast}
\alias{CapitalForecast}
\title{This function outputs a time
series capital forecast from time series inputs PPNR, AFS
and LLL forecasts.}
\usage{
CapitalForecast(position_data, model_coefficients, macro_forecasts)
}
\arguments{
\item{position_data}{A data frame of a particular
bank's present balance sheet}

\item{model_coefficients}{A data frame of arima calibration
coefficients computed from past history}

\item{macro_forecasts}{A time series of either basic, adverse, or severely
adverse macroeconomic forecasts#'}
}
\value{
object of class MTS
}
\description{
This function outputs a time
series capital forecast from time series inputs PPNR, AFS
and LLL forecasts.
}
\author{
David Karapetyan
}

