% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ppnr_forecast.R
\name{PPNRForecast}
\alias{PPNRForecast}
\title{This function outputs a PPNR
forecast from input files for bank position data,
macroeconomic forecasts , and Arima model coefficients.}
\usage{
PPNRForecast(position_data, model_coefficients, macro_forecasts)
}
\arguments{
\item{position_data}{A matrix of a particular bank's
present balance sheet}

\item{model_coefficients}{A matrix of arima calibration
coefficients computed from past history}

\item{macro_forecasts}{A matrix of either basic,
adverse, or severely adverse macroeconomic forecasts}
}
\value{
object of class MTS
}
\description{
This function outputs a PPNR
forecast from input files for bank position data,
macroeconomic forecasts , and Arima model coefficients.
}
\author{
David Karapetyan
}

