% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/asset_forecast.R
\name{RevenueForecast}
\alias{RevenueForecast}
\title{This function outputs a PPNR asset
forecast from an input file for bank position data.}
\usage{
RevenueForecast(position_data, macro_forecasts)
}
\arguments{
\item{position_data}{A matrix of a particular bank's
present balance sheet}

\item{macro_forecasts}{A matrix of either basic,
adverse, or severely adverse macroeconomic forecasts}
}
\value{
object of class MTS
}
\description{
This function outputs a PPNR asset
forecast from an input file for bank position data.
}
\author{
David Karapetyan
}

